Update：修改tar压缩/解压缩示例

+++ src/main/java/com/gonghf/file/compression/tar/TarUtil.java
@@ -4,6 +4,7 @@ import com.gonghf.file.compression.commom.FileUtil;
 import org.apache.commons.compress.archivers.tar.TarArchiveEntry;
 import org.apache.commons.compress.archivers.tar.TarArchiveInputStream;
 import org.apache.commons.compress.archivers.tar.TarArchiveOutputStream;
+
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
@@ -11,21 +12,30 @@ import java.util.List;
 public class TarUtil {
 
     /**
-     * @param dir 需要打包的目录或文件
+     * @param dir 需要压缩的目录或文件
      * @return 打包后tar包的地址
      * @throws Exception
      */
     public static String compression(String dir) throws Exception {
-        List<File> dirFile = FileUtil.findDirFile(dir);
         File file = new File(dir);
-        String parentPath = file.getParentFile().getAbsolutePath();
+        if (!file.exists()) {
+            return null;
+        }
+        //获取目录下所有文件
+        List<File> dirFiles = FileUtil.findDirFile(dir);
+
         String fileName = file.getName();
+        String parentPath = file.getParentFile().getAbsolutePath();
+        //获取压缩后的文件路径
         String resultFilePath = parentPath + File.separator + fileName + ".tar";
+
+        //如果压缩后的文件存在，先进行删除
         File resultFile = new File(resultFilePath);
         if (resultFile.exists()) {
             resultFile.delete();
         }
-        boolean compression = tarCompression(dir,dirFile.toArray(new File[dirFile.size()]), resultFilePath);
+
+        boolean compression = tarCompression(dir, dirFiles, resultFilePath);
         if (compression) {
             return resultFilePath;
         } else {
@@ -35,13 +45,15 @@ public class TarUtil {
 
     /**
      * @param tarFile tar包地址
-     * @param outdir 解压缩的目录
+     * @param outdir  解压缩的目录
      * @return 解压后所有文件的集合
      * @throws Exception
      */
     public static File[] decompression(String tarFile, String outdir) throws Exception {
-        ArrayList<File> filelist = new ArrayList<>();
-        boolean decompression = tarDecompression(tarFile, outdir, filelist);
+        //解压后的文件集合
+        List<File> filelist = new ArrayList<>();
+        //解压文件
+        boolean decompression = tarDecompression(tarFile, filelist, outdir);
         if (decompression) {
             return filelist.toArray(new File[filelist.size()]);
         } else {
@@ -52,26 +64,27 @@ public class TarUtil {
     /**
      * tar打包压缩
      *
-     * @param filesArray     要压缩的文件的全路径(数组)
+     * @param fileslist      要压缩的文件的全路径(集合)
      * @param resultFilePath 压缩后的文件全文件名(.tar)
      * @throws Exception
      * @DATE 2018年9月25日 下午12:39:28
      */
-    private static boolean tarCompression(String basePath,File[] filesArray, String resultFilePath) throws Exception {
+    private static boolean tarCompression(String basePath, List<File> fileslist, String resultFilePath) throws Exception {
         System.out.println(" tarCompression -> Compression start!");
         FileOutputStream fos = null;
         TarArchiveOutputStream taos = null;
+        String baseFilePath = new File(basePath).getAbsolutePath();
         try {
             fos = new FileOutputStream(new File(resultFilePath));
             taos = new TarArchiveOutputStream(fos);
-            for (File file : filesArray) {
+            for (File file : fileslist) {
                 BufferedInputStream bis = null;
                 FileInputStream fis = null;
                 try {
                     TarArchiveEntry tae = new TarArchiveEntry(file);
                     tae.setSize(file.length());
-                    String tarPath = file.getAbsolutePath().replaceAll(basePath, "");
-                    tae.setName(file.getAbsolutePath());
+                    String tarPath = file.getAbsolutePath().replace(baseFilePath,"");
+                    tae.setName(tarPath);
                     taos.putArchiveEntry(tae);
                     System.out.println("  compression file -> " + tae.getName());
                     fis = new FileInputStream(file);
@@ -107,12 +120,15 @@ public class TarUtil {
      * @throws Exception
      * @DATE 2018年9月25日 下午12:39:43
      */
-    private static boolean tarDecompression(String decompressFilePath, String resultDirPath, List<File> fileList) throws Exception {
+    private static boolean tarDecompression(String decompressFilePath, List<File> fileList, String resultDirPath) throws Exception {
         System.out.println(" tarDecompression -> Decompression start!");
+        File file = new File(decompressFilePath);
+        if(!file.exists()){
+            return false;
+        }
         TarArchiveInputStream tais = null;
         FileInputStream fis = null;
         try {
-            File file = new File(decompressFilePath);
             fis = new FileInputStream(file);
             tais = new TarArchiveInputStream(fis);
             TarArchiveEntry tae = null;
@@ -124,7 +140,7 @@ public class TarUtil {
                     String dir = resultDirPath + File.separator + tae.getName();// tar档中文件
                     File dirFile = new File(dir);
                     fileList.add(dirFile);
-                    if(!dirFile.getParentFile().exists()){
+                    if (!dirFile.getParentFile().exists()) {
                         dirFile.getParentFile().mkdirs();
                     }
                     fos = new FileOutputStream(dirFile);
